pyproject.toml
[tool.poetry]
name = "price-parser"
version = "0.3.0"
description = ""
authors = ["Nyaaa <1457047+Nyaaa@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.11.2"
python-dotenv = "^1.0.0"
lxml = "^4.9.2"
aiohttp = "^3.8.4"
coloredlogs = "^15.0.1"
aiohttp-retry = "^2.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
